add atomic design to components, change service to provider and create App bootstrap component
