--- src/components/FilesDragDrop/index.js
@@ -1,1 +1,0 @@
-export { default } from './FilesDragDrop';
